commit by xqs

+++ src/main/java/top/philxin/model/WxCartModel/CheckOutVo.java
@@ -31,7 +31,7 @@ public class CheckOutVo {
 
     BigDecimal couponPrice;
 
-    int availableCouponLength;
+    Short availableCouponLength;
 
     int couponId;
